Consolidate artifacts queries into fragment
