+++ src/fragments/artifacts.graphql
@@ -1,0 +1,18 @@
+fragment Artifacts on ArtifactsConnection {
+  pageInfo {
+    hasNextPage
+    hasPreviousPage
+    cursor
+    previousCursor
+    nextCursor
+  }
+
+  edges {
+    node {
+      name
+      contentType
+      url
+      isPublicLog
+    }
+  }
+}
