+++ src/utils/constants.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const ARTIFACTS_PAGE_SIZE = 10;
