+++ src/views/Tasks/ViewTask/pageArtifacts.graphql
@@ -1,24 +1,11 @@
+#import "../../../fragments/artifacts.graphql"
+
 query PageArtifacts(
   $taskId: ID!,
   $runId: Int!,
   $connection: PageConnection!
 ) {
   artifacts(taskId: $taskId, runId: $runId, connection: $connection) {
-    pageInfo {
-      hasNextPage
-      hasPreviousPage
-      cursor
-      previousCursor
-      nextCursor
-    }
-
-    edges {
-      node {
-        name
-        contentType
-        url
-        isPublicLog
-      }
-    }
+    ...Artifacts
   }
 }
